abstract out the meshnode
